Do not set the terminal to raw mode if it isn't actually a terminal
It doesn't make sense to try to set a terminal as a raw terminal if it
isn't actually a terminal. Worse, this may break tests that assume they
can simply provide a pipe to a subprocess and expect the terminal will
be unchanged.

See this PR, which needs this patch:
https://github.com/sharkdp/bat/pull/2631